Revert "convert field variable"
This reverts commit d09c48aade1faa21376c2b74f07e88c99440e05e.
